search_history sameSearchUpdate
